refactor(bayn): modularize reference replay
